+++ scripts/wv2gtk-build.sh
@@ -76,6 +76,9 @@ CAPTURE_VALA=(
 	lib/webview2gtk/NetworkProxySettings.vala
 	lib/webview2gtk/CookieManager.vala
 	lib/webview2gtk/URIRequest.vala
+	lib/webview2gtk/PolicyDecision.vala
+	lib/webview2gtk/URIResponse.vala
+	lib/webview2gtk/ResponsePolicyDecision.vala
 	lib/webview2gtk/WebResource.vala
 	lib/webview2gtk/Download.vala
 	lib/webview2gtk/NetworkSession.vala